+++ src/include/reb-host.h
@@ -57,7 +57,6 @@
 #endif
 
 #include "reb-ext.h"        // includes reb-defs.h
-#include "reb-args.h"
 #include "reb-device.h"
 #include "reb-file.h"
 #include "reb-event.h"
