+++ src/include/sys-core.h
@@ -207,7 +207,6 @@ extern void reb_qsort_r(void *a, size_t n, size_t es, void *thunk, cmp_t *cmp);
 #endif
 
 #include "reb-defs.h"
-#include "reb-args.h"
 
 #include "reb-device.h"
 #include "reb-types.h"
